Moving forms items into their own folders
